Separate the content and tab bar components
Create a new component called ContentTabs and rename the
ContentTabbedBar to TabbedNavbar.

Move all the middle section in a row container.
